# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API for Jumyste application",
        "title": "Jumyste App API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api",
    "paths": {
...
        "/vacancies": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieves a list of all vacancies",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vacancies"
                ],
                "summary": "Get all vacancies",
                "responses": {
                    "200": {
                        "description": "List of vacancies",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/jumyste-app-backend_internal_entity.Vacancy"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to fetch vacancies",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows an employer to create a new vacancy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vacancies"
                ],
                "summary": "Create a new vacancy",
                "parameters": [
                    {
                        "description": "Vacancy details",
                        "name": "vacancy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.CreateVacancyRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Vacancy successfully created",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_entity.Vacancy"
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to create vacancy",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/vacancies/company": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve all vacancies for the company of the current HR",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vacancies"
                ],
                "summary": "Get vacancies by company ID",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/jumyste-app-backend_internal_entity.Vacancy"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve vacancies",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/vacancies/my": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a list of vacancies created by the currently authenticated HR user",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vacancies"
                ],
                "summary": "Get vacancies created by the authenticated HR",
                "responses": {
                    "200": {
                        "description": "List of vacancies",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/jumyste-app-backend_internal_entity.Vacancy"
                            }
                        }
                    },
                    "500": {
                        "description": "Failed to retrieve vacancies",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/vacancies/search": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows searching for vacancies based on various filters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vacancies"
                ],
                "summary": "Search for vacancies",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Search query",
                        "name": "query",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Employment type filter",
                        "name": "employment_type",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Work format filter",
                        "name": "work_format",
                        "in": "query"
                    },
                    {
                        "type": "array",
                        "items": {
                            "type": "string"
                        },
                        "collectionFormat": "multi",
                        "description": "Skills filter",
                        "name": "skills",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of matching vacancies",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/jumyste-app-backend_internal_entity.Vacancy"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid search parameters",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to search vacancies",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/vacancies/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Retrieve a specific vacancy by its ID",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vacancies"
                ],
                "summary": "Get vacancy by ID",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vacancy ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_entity.Vacancy"
                        }
                    },
                    "400": {
                        "description": "Invalid vacancy ID",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Vacancy not found",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows an employer to update their own vacancy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vacancies"
                ],
                "summary": "Update an existing vacancy",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vacancy ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated vacancy details",
                        "name": "vacancy",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.UpdateVacancyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vacancy updated successfully",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid input or vacancy ID",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "User does not own the vacancy",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Vacancy not found",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to update vacancy",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Allows an employer to delete their own vacancy",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Vacancies"
                ],
                "summary": "Delete a vacancy",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Vacancy ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Vacancy deleted successfully",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid vacancy ID",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "User does not own the vacancy",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Vacancy not found",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Failed to delete vacancy",
                        "schema": {
                            "$ref": "#/definitions/jumyste-app-backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "jumyste-app-backend_internal_dto.CreateVacancyRequest": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string",
                    "example": "IT"
                },
                "description": {
                    "type": "string",
                    "example": "\u003ch3\u003e\u003cstrong\u003e\u003cem\u003e\u003cs\u003eHello\u003c/s\u003e\u003c/em\u003e\u003c/strong\u003e\u003c/h3\u003e"
                },
                "employment_type": {
                    "type": "string",
                    "example": "Полная занятость"
                },
                "experience": {
                    "type": "string",
                    "example": "Без опыта"
                },
                "location": {
                    "type": "string",
                    "example": "Almaty"
                },
                "salary_max": {
                    "type": "integer",
                    "example": 909000
                },
                "salary_min": {
                    "type": "integer",
                    "example": 253000
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    },
                    "example": [
                        "[\"Python\"",
                        "\"Node.js\"]"
                    ]
                },
                "title": {
                    "type": "string",
                    "example": "Frontend developer"
                },
                "work_format": {
                    "type": "string",
                    "example": "Гибрид"
                }
            }
        },
        "jumyste-app-backend_internal_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "Invalid input"
                }
            }
        },
...
        "jumyste-app-backend_internal_dto.SuccessResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "User registered successfully"
                }
            }
        },
        "jumyste-app-backend_internal_dto.UpdateVacancyRequest": {
            "type": "object",
            "required": [
                "category",
                "description",
                "employment_type",
                "experience",
                "location",
                "salary_max",
                "salary_min",
                "skills",
                "title",
                "work_format"
            ],
            "properties": {
                "category": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "employment_type": {
                    "type": "string"
                },
                "experience": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "salary_max": {
                    "type": "integer"
                },
                "salary_min": {
                    "type": "integer"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "work_format": {
                    "type": "string"
                }
            }
        },
...
        "jumyste-app-backend_internal_entity.Vacancy": {
            "type": "object",
            "properties": {
                "category": {
                    "type": "string"
                },
                "company_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "integer"
                },
                "description": {
                    "type": "string"
                },
                "employment_type": {
                    "type": "string"
                },
                "experience": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "location": {
                    "type": "string"
                },
                "salary_max": {
                    "type": "integer"
                },
                "salary_min": {
                    "type": "integer"
                },
                "skills": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "work_format": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}